requirements.txt
django
djangorestframework
markdown
pylint_django
coverage
django-filter
